Added a different setting to the template creation step
Instructions for configuring Sensitivity Labels in Azure AD directs you to this page via three other pages in the event that your settings are not yet created. To create the settings this page has you add a Usage Guideline, but the purpose is to create any value that makes the setting and not just Usage Guidelines. Added this info to the document so that the actual steps needed are not adding UsageGuidelinesUrl when actually the admin might want EnableMIPLabels